src/serving/pyproject.toml
[project]
name = "serving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "mlflow>=3.8.1",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "python-multipart>=0.0.21",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "uvicorn>=0.40.0",
]
